Project Setup (Camera Issue)
I was testing pulling from the master and working off of the project. Between the old project (different name on my local) and this project, the camera broke. Had to update the project settings to make default texture filter = nearest and zoom in on the player camera by 3x. Hopefully these changes stick.