New Assets and Scenes
Added 2 new asset tilesets.
Created the forest world scene.
Created room of choice scene.
started working on transition between scenes. It is a little buggy, need to look into Autoload (Singletons) to store loading positions of Morgan when going in and out of different areas.
Created a global group called Player and assigned Morgan to it so it can be used for signals.